FIles now changes when comands is called
DeleteControl and SyncFiles changed
Now you can delete both files or content of both files just calling the functions from cmd
SyncFiles automatically copy all original file content when is called

+++ DeleteControl.cs
@@ -6,28 +6,56 @@ namespace VEEAM_Test_Task
     internal class DeleteControl
     {
 
-        public static void deleteArchive(string test)
+        public static void deleteArchive(string originalFilePathing, string copyFilePathing = "")
         {
             Console.WriteLine("Delete archive?");
             Console.WriteLine("1-Yes");
-            Console.WriteLine("2-Back to main menu");
+            Console.WriteLine("2-Delete Content");
+            Console.WriteLine("3-Back to main menu");
             string deleteOptions = Console.ReadLine();
 
-            string archivePath = archivePathing(test);
+            string archivePath = archivePathing(originalFilePathing);
+            string fileCopyArchivePath = archivePathing(copyFilePathing);
 
+            var archives = Directory.GetFiles(archivePath);
+            var copyArchives = Directory.GetFiles(fileCopyArchivePath);
 
 
             switch (deleteOptions)
             {
                 case "1":
-                    if (Directory.Exists(archivePath))
+                    if (Directory.Exists(archivePath) && Directory.Exists(fileCopyArchivePath))
                     {
                         Directory.Delete(archivePath, true);
+                        Directory.Delete(fileCopyArchivePath, true);
                     }
                     MainSelector.processChoices();
                     break;
 
                 case "2":
+                    try
+                    {
+                        foreach (string loopFile in archives)
+                        {
+                            string deleteContent = Path.GetFileName(loopFile);
+                            File.Delete(loopFile);
+
+                            foreach (string loopCopyFile in copyArchives)
+                            {
+                                string pathingCopyDelete = Path.GetFileName(loopCopyFile);
+                                File.Delete(loopCopyFile);
+                            }
+                        }
+
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine($"error try again ++ {ex}");
+                    }
+                    MainSelector.processChoices();
+                    break;
+
+                case "3":
                     MainSelector.processChoices();
                     break;
